refactor: improve file structure organization for types
Rearrange the file structure of types to enhance organization and maintenance.
Consolidate all types into /lib/types/index as a barrel file, serving as a
single source for imports. To maintain consistency, a new eslint rule is added,
disallowing direct imports from /lib/types/bases/*.
