+++ components/icons/svgLogo.tsx
@@ -1,6 +1,6 @@
 import { DATA_SVG_LOGO } from '@collections/svgLogo';
 import { SVG_LOGO } from '@constAssertions/data';
-import { TypesOptionsSvg } from '@lib/types/typesOptions';
+import { TypesOptionsSvg } from '@lib/types/options';
 
 type Props = { options: TypesOptionsSvg } & { svgLogo: SVG_LOGO };
