+++ components/todos/todo/todoItem.tsx
@@ -6,9 +6,10 @@ import { useTodoCompleteItem } from '@hooks/todos';
 import { CheckBox } from '@inputs/checkbox';
 import { TypesTodo } from '@lib/types';
 import { optionsSvgPriorityUrgent, optionsSvgPriorityImportant, optionsSvgCalendarDueDate } from '@options/svg';
-import { atomCatch, classNames } from '@stateLogics/utils';
+import { classNames } from '@stateLogics/utils';
 import { selectorSessionTodoItem } from '@states/atomEffects/todos';
 import { selectorSelectedQueryLabels } from '@states/labels';
+import { atomCatch } from '@states/misc';
 import { format } from 'date-fns';
 import { Fragment as CheckBoxFragment, Fragment as TodoItemFragment } from 'react';
 import { useRecoilValue } from 'recoil';
