+++ components/ui/buttons/button/index.tsx
@@ -1,4 +1,4 @@
-import { TypesOptionsButton } from '@lib/types/typesOptions';
+import { TypesOptionsButton } from '@lib/types/options';
 import { Types } from 'lib/types';
 import dynamic from 'next/dynamic';
 import { forwardRef, useState } from 'react';
