+++ components/ui/buttons/button/svgLogoButton.tsx
@@ -1,9 +1,9 @@
 import { TypesSvgLogo } from '@lib/types';
-import { TypesOptionsSvg } from '@lib/types/typesOptions';
 import { signIn } from 'next-auth/react';
 import { useRouter } from 'next-router-mock';
 import { Button } from '.';
 import { DATA_SVG_LOGO } from '@collections/svgLogo';
+import { TypesOptionsSvg } from '@lib/types/options';
 
 type Props = { options?: TypesOptionsSvg };
