+++ components/ui/buttons/iconButton/priorityButton.tsx
@@ -3,7 +3,7 @@ import { SvgIcon } from '@components/icons/svgIcon';
 import { PRIORITY_LEVEL } from '@constAssertions/misc';
 import { ICON_FLAG, ICON_FLAG_FILL, ICON_LABEL_IMPORTANT, ICON_LABEL_IMPORTANT_FILL } from '@data/materialSymbols';
 import { Types } from '@lib/types';
-import { TypesOptionsPriority } from '@lib/types/typesOptions';
+import { TypesOptionsPriority } from '@lib/types/options';
 import { classNames } from '@stateLogics/utils';
 import { atomSelectorTodoItem, selectorSessionTodoItem } from '@states/atomEffects/todos';
 import { atomTodoNew } from '@states/todos';
