+++ components/ui/dropdowns/v1/dropdown/index.tsx
@@ -1,6 +1,5 @@
 import { STYLE_HOVER_SLATE_DARK } from '@data/stylePreset';
 import { Menu, Transition } from '@headlessui/react';
-import { TypesOptionsDropdown } from '@lib/types/typesOptions';
 import { atomOnBlur } from '@states/focus';
 import { SvgIcon } from 'components/icons/svgIcon';
 import { Types } from 'lib/types';
@@ -11,6 +10,7 @@ import { useSetRecoilState } from 'recoil';
 import { ConditionalPortal } from './conditionalPortal';
 import { DisableScrollEffect } from '@effects/disableScrollEffect';
 import { classNames } from '@stateLogics/utils';
+import { TypesOptionsDropdown } from '@lib/types/options';
 const Tooltip = dynamic(() => import('@tooltips/tooltips').then((mod) => mod.Tooltip));
 
 type Props = { options: TypesOptionsDropdown } & Partial<
