+++ components/ui/dropdowns/v1/labelItemDropdown.tsx
@@ -1,5 +1,4 @@
 import { ICON_DELETE, ICON_EDIT_NOTE } from '@data/materialSymbols';
-import { TypesOptionsDropdown } from '@lib/types/typesOptions';
 import { Types } from 'lib/types';
 import { Dropdown } from './dropdown';
 import { DropdownMenuItem } from './dropdown/dropdownMenuItem';
@@ -7,6 +6,7 @@ import { optionsDropdownLabelItem } from '@options/misc';
 import { ActiveDropdownMenuItemEffect } from '@effects/activeDropdownMenuItemEffect';
 import { useLabelRemoveItem } from '@hooks/labels';
 import { useLabelModalStateOpen } from '@hooks/modals';
+import { TypesOptionsDropdown } from '@lib/types/options';
 
 type Props = { options: TypesOptionsDropdown } & Pick<Types, 'label'> & Partial<Pick<Types, 'menuContentOnClose'>>;
