+++ components/ui/dropdowns/v2/dropdown/menuItems.tsx
@@ -2,7 +2,7 @@ import { ConditionalPortal } from '@dropdowns/v1/dropdown/conditionalPortal';
 import { DisableScrollEffect } from '@effects/disableScrollEffect';
 import { Menu, Transition } from '@headlessui/react';
 import { Types } from '@lib/types';
-import { TypesOptionsDropdown } from '@lib/types/typesOptions';
+import { TypesOptionsDropdown } from '@lib/types/options';
 import { classNames } from '@stateLogics/utils';
 import { useState } from 'react';
 import { usePopper } from 'react-popper';
