+++ components/ui/pseudoButtons/pseudoIconButton/index.tsx
@@ -1,5 +1,5 @@
 import { Types } from '@lib/types';
-import { TypesOptionsButton } from '@lib/types/typesOptions';
+import { TypesOptionsButton } from '@lib/types/options';
 import { classNames } from '@stateLogics/utils';
 import dynamic from 'next/dynamic';
 import { Fragment as HeaderContentsFragment } from 'react';
