+++ lib/data/options/button.tsx
@@ -20,7 +20,7 @@ import {
   STYLE_BUTTON_NORMAL_WHITE,
   STYLE_HOVER_ENABLED_SLATE_DARK,
 } from '@data/stylePreset';
-import { TypesOptionsButton } from '@lib/types/typesOptions';
+import { TypesOptionsButton } from '@lib/types/options';
 import { classNames } from '@stateLogics/utils';
 
 // global
