+++ lib/stateLogics/effects/KeysWithNavigateEffect.tsx
@@ -2,9 +2,9 @@ import { CATCH, FOCUS } from '@constAssertions/misc';
 import { useFocusState } from '@hooks/focus';
 import { useKeyWithNavigate } from '@hooks/keybindings';
 import { Types } from '@lib/types';
-import { atomCatch } from '@stateLogics/utils';
 import { selectorSessionTodoItem } from '@states/atomEffects/todos';
 import { atomCurrentFocus, atomOnBlur } from '@states/focus';
+import { atomCatch } from '@states/misc';
 import { atomTodoModalOpen } from '@states/modals';
 import { useEffect } from 'react';
 import { RecoilValue, useRecoilCallback, useRecoilValue } from 'recoil';
