+++ lib/stateLogics/effects/sidebarMobileResetEffect.tsx
@@ -1,10 +1,9 @@
-import { atomMediaQuery } from '@states/misc';
+import { atomCatch, atomMediaQuery } from '@states/misc';
 import { useEffect } from 'react';
 import { RecoilValue, useRecoilCallback, useRecoilValue } from 'recoil';
 import { atomSidebarOpenMobile } from '@states/layouts';
 import { CATCH } from '@constAssertions/misc';
 import { BREAKPOINT } from '@constAssertions/ui';
-import { atomCatch } from '@stateLogics/utils';
 
 export const SidebarMobileResetEffect = () => {
   const isTodoModalOpen = useRecoilValue(atomCatch(CATCH['todoModal']) || CATCH['confirmModal']);
