+++ lib/stateLogics/utils/index.tsx
@@ -1,22 +1,8 @@
-import { CATCH } from '@constAssertions/misc';
 import { STORAGE_KEY } from '@constAssertions/storage';
 import { render, RenderOptions } from '@testing-library/react';
 import React, { FC, ReactElement } from 'react';
-import { atom, atomFamily, RecoilRoot } from 'recoil';
+import { RecoilRoot } from 'recoil';
 import validator from 'validator';
-
-/**
- * Atoms
- */
-export const atomCatch = atomFamily<boolean, CATCH>({
-  key: 'atomCatch',
-  default: false,
-});
-
-export const atomDisableScroll = atom<boolean>({
-  key: 'atomDisableScroll',
-  default: false,
-});
 
 /**
  * Utils
